pyproject.toml
[project]
name = "radio-g"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "black>=25.12.0",
    "confluent-kafka>=2.12.2",
    "coverage[toml]>=7.13.0",
    "elevenlabs>=2.27.0",
    "fastapi>=0.127.0",
    "feedparser>=6.0.12",
    "flake8>=7.3.0",
    "google-genai>=1.56.0",
    "isort>=7.0.0",
    "lxml[html-clean]>=6.0.2",
    "newspaper3k>=0.2.8",
    "psutil>=7.2.0",
    "pydantic>=2.12.5",
    "pylint>=4.0.4",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-clarity>=1.0.1",
    "pytest-cov>=7.0.0",
    "pytest-fixtures>=0.1.0",
    "pytest-html>=4.1.1",
    "pytest-mock>=3.15.1",
    "pytest-timeout>=2.4.0",
    "pytest-xdist>=3.8.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "python-json-logger>=4.0.0",
    "python-multipart>=0.0.21",
    "pyyaml>=6.0.3",
    "redis>=7.1.0",
    "requests>=2.32.5",
    "schedule>=1.2.2",
    "urllib3>=2.6.2",
    "uvicorn[standard]>=0.40.0",
]
